initial ExtractionService implementation
