+++ packages/server/src/main/java/software/uncharted/terarium/hmiserver/controller/knowledge/KnowledgeController.java
@@ -71,6 +71,7 @@ import software.uncharted.terarium.hmiserver.proxies.mit.MitProxy;
 import software.uncharted.terarium.hmiserver.proxies.skema.SkemaUnifiedProxy;
 import software.uncharted.terarium.hmiserver.proxies.skema.SkemaUnifiedProxy.IntegratedTextExtractionsBody;
 import software.uncharted.terarium.hmiserver.security.Roles;
+import software.uncharted.terarium.hmiserver.service.ExtractionService;
 import software.uncharted.terarium.hmiserver.service.data.CodeService;
 import software.uncharted.terarium.hmiserver.service.data.DatasetService;
 import software.uncharted.terarium.hmiserver.service.data.DocumentAssetService;
@@ -101,6 +102,8 @@ public class KnowledgeController {
 	final ProvenanceSearchService provenanceSearchService;
 
 	final CodeService codeService;
+
+	final ExtractionService extractionService;
 
 	@Value("${mit-openai-api-key:}")
 	String MIT_OPENAI_API_KEY;
@@ -212,9 +215,6 @@ public class KnowledgeController {
 	 * Transform source code to AMR
 	 *
 	 * @param codeId       (String): id of the code artifact
-	 * @param name         (String): the name to set on the newly created model
-	 * @param description  (String): the description to set on the newly created
-	 *                     model
 	 * @param dynamicsOnly (Boolean): whether to only run the amr extraction over
 	 *                     specified dynamics from the code object in TDS
 	 * @param llmAssisted  (Boolean): whether amr extraction is llm assisted
@@ -768,207 +768,12 @@ public class KnowledgeController {
 		}
 	}
 
-	public HttpEntity zipEntryToHttpEntity(final ZipInputStream zipInputStream) throws IOException {
-		final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-		final byte[] buffer = new byte[1024];
-		int len;
-		while ((len = zipInputStream.read(buffer)) > 0) {
-			byteArrayOutputStream.write(buffer, 0, len);
-		}
-
-		return new ByteArrayEntity(byteArrayOutputStream.toByteArray());
-	}
-
 	@PostMapping("/pdf-to-cosmos")
 	@Secured(Roles.USER)
 	public ResponseEntity<Void> postPDFToCosmos(
 			@RequestParam("document-id") final UUID documentId) {
-
-		new Runnable() {
-
-			@Override
-			public void run() {
-
-			}
-		};
-		Return 202;
-
-		try {
-			DocumentAsset document = documentService.getAsset(documentId).orElseThrow();
-
-			if (document.getFileNames().isEmpty()) {
-				throw new RuntimeException("No files found on document");
-			}
-
-			final String filename = document.getFileNames().get(0);
-
-			final byte[] documentContents = documentService.fetchFileAsBytes(documentId, filename).orElseThrow();
-
-			final ByteMultipartFile documentFile = new ByteMultipartFile(documentContents, filename,
-					"application/pdf");
-
-			final boolean compressImages = false;
-			final boolean forceRun = false;
-			final ResponseEntity<JsonNode> extractionResp = extractionProxy.processPdfExtraction(compressImages,
-					!forceRun,
-					documentFile);
-
-			final JsonNode body = extractionResp.getBody();
-			final UUID jobId = UUID.fromString(body.get("job_id").asText());
-
-			final int POLLING_INTERVAL_SECONDS = 5;
-			final int MAX_EXECUTION_TIME_SECONDS = 600;
-			final int MAX_ITERATIONS = MAX_EXECUTION_TIME_SECONDS / POLLING_INTERVAL_SECONDS;
-
-			boolean jobDone = false;
-			for (int i = 0; i < MAX_ITERATIONS; i++) {
-
-				final ResponseEntity<JsonNode> statusResp = extractionProxy.status(jobId);
-				if (!statusResp.getStatusCode().is2xxSuccessful()) {
-					throw new RuntimeException("Unable to poll status endpoint");
-				}
-
-				final JsonNode statusData = statusResp.getBody();
-				if (!statusData.get("error").isNull()) {
-					throw new RuntimeException("Extraction job failed: " + statusData.has("error"));
-				}
-
-				log.info("Polled status endpoint {} times:\n{}", i + 1, statusData);
-				jobDone = statusData.get("error").asBoolean() || statusData.get("job_completed").asBoolean();
-				if (jobDone) {
-					break;
-				}
-				Thread.sleep(POLLING_INTERVAL_SECONDS * 1000);
-			}
-
-			if (!jobDone) {
-				throw new RuntimeException("Extraction job did not complete within the expected time");
-			}
-
-			final ResponseEntity<byte[]> zipFileResp = extractionProxy.result(jobId);
-			if (!zipFileResp.getStatusCode().is2xxSuccessful()) {
-				throw new RuntimeException("Unable to fetch the extraction result");
-			}
-
-			final String zipFileName = documentId + "_cosmos.zip";
-			documentService.uploadFile(documentId, zipFileName, new ByteArrayEntity(zipFileResp.getBody()));
-
-			document.getFileNames().add(zipFileName);
-
-			// Open the zipfile and extract the contents
-
-			final Map<String, HttpEntity> fileMap = new HashMap<>();
-			try {
-				final ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(zipFileResp.getBody());
-				final ZipInputStream zipInputStream = new ZipInputStream(byteArrayInputStream);
-
-				ZipEntry entry = zipInputStream.getNextEntry();
-				while (entry != null) {
-
-					fileMap.put(entry.getName(), zipEntryToHttpEntity(zipInputStream));
-
-					entry = zipInputStream.getNextEntry();
-				}
-
-				zipInputStream.closeEntry();
-				zipInputStream.close();
-			} catch (final IOException e) {
-				throw new RuntimeException("Unable to extract the contents of the zip file", e);
-			}
-
-			final ResponseEntity<JsonNode> textResp = extractionProxy.text(jobId);
-			if (!textResp.getStatusCode().is2xxSuccessful()) {
-				throw new RuntimeException("Unable to fetch the text extractions");
-			}
-
-			// clear existing assets
-			document.setAssets(new ArrayList<>());
-
-			for (final ExtractionAssetType extractionType : ExtractionAssetType.values()) {
-				final ResponseEntity<JsonNode> response = extractionProxy.extraction(jobId,
-						extractionType.toStringPlural());
-				log.info(" {} response status: {}", extractionType, response.getStatusCode());
-				if (!response.getStatusCode().is2xxSuccessful()) {
-					log.warn("Unable to fetch the {} extractions", extractionType);
-					continue;
-				}
-
-				for (final JsonNode record : response.getBody()) {
-
-					String fileName = "";
-					if (record.has("img_pth")) {
-
-						final String path = record.get("img_pth").asText();
-						fileName = path.substring(path.lastIndexOf("/") + 1);
-
-						if (fileMap.containsKey(fileName)) {
-							log.warn("Unable to find file {} in zipfile", fileName);
-						}
-
-						final HttpEntity file = fileMap.get(fileName);
-
-						documentService.uploadFile(documentId, fileName, file);
-
-					} else {
-						log.warn("No img_pth found in record: {}", record);
-					}
-
-					final DocumentExtraction extraction = new DocumentExtraction();
-					extraction.setFileName(fileName);
-					extraction.setAssetType(extractionType);
-					extraction.setMetadata(mapper.convertValue(record, Map.class));
-
-					document.getAssets().add(extraction);
-				}
-			}
-
-			String responseText = "";
-			for (final JsonNode record : textResp.getBody()) {
-				if (record.has("content")) {
-					responseText += record.get("content").asText() + "\n";
-				} else {
-					log.warn("No content found in record: {}", record);
-				}
-			}
-
-			document.setText(responseText);
-
-			// update the document
-			document = documentService.updateAsset(document).orElseThrow();
-
-			if (document.getText() == null || document.getText().isEmpty()) {
-				log.warn("Document {} has no text to send", documentId);
-				throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Document has no text");
-			}
-
-			// check for input length
-			if (document.getText().length() > 600000) {
-				log.warn("Document {} text too long for GoLLM model card task", documentId);
-				throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Document text is too long");
-			}
-
-			final ModelCardResponseHandler.Input input = new ModelCardResponseHandler.Input();
-			input.setResearchPaper(document.getText());
-
-			// Create the task
-			final TaskRequest req = new TaskRequest();
-			req.setType(TaskRequest.TaskType.GOLLM);
-			req.setScript(ModelCardResponseHandler.NAME);
-			req.setInput(mapper.writeValueAsBytes(input));
-
-			final ModelCardResponseHandler.Properties props = new ModelCardResponseHandler.Properties();
-			props.setDocumentId(documentId);
-			req.setAdditionalProperties(props);
-
-			return ResponseEntity.accepted().build();
-
-		} catch (final Exception e) {
-			final String error = "Unable to extract pdf";
-			log.error(error, e);
-			throw new ResponseStatusException(
-					org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR,
-					error);
-		}
+		extractionService.extractPDF(documentId);
+		return ResponseEntity.accepted().build();
 	}
 
 }
