+++ packages/server/src/main/java/software/uncharted/terarium/hmiserver/service/ExtractionService.java
@@ -1,0 +1,226 @@
+package software.uncharted.terarium.hmiserver.service;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.http.HttpEntity;
+import org.apache.http.entity.ByteArrayEntity;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+import software.uncharted.terarium.hmiserver.models.dataservice.document.DocumentAsset;
+import software.uncharted.terarium.hmiserver.models.dataservice.document.DocumentExtraction;
+import software.uncharted.terarium.hmiserver.models.dataservice.document.ExtractionAssetType;
+import software.uncharted.terarium.hmiserver.models.task.TaskRequest;
+import software.uncharted.terarium.hmiserver.proxies.documentservice.ExtractionProxy;
+import software.uncharted.terarium.hmiserver.service.data.DocumentAssetService;
+import software.uncharted.terarium.hmiserver.service.tasks.ModelCardResponseHandler;
+import software.uncharted.terarium.hmiserver.utils.ByteMultipartFile;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class ExtractionService {
+    final DocumentAssetService documentService;
+    final ExtractionProxy extractionProxy;
+    final ObjectMapper objectMapper;
+
+    public void extractPDF(UUID documentId) {
+        try {
+            DocumentAsset document = documentService.getAsset(documentId).orElseThrow();
+
+            if (document.getFileNames().isEmpty()) {
+                throw new RuntimeException("No files found on document");
+            }
+
+            final String filename = document.getFileNames().get(0);
+
+            final byte[] documentContents = documentService.fetchFileAsBytes(documentId, filename).orElseThrow();
+
+            final ByteMultipartFile documentFile = new ByteMultipartFile(documentContents, filename,
+                    "application/pdf");
+
+            final boolean compressImages = false;
+            final boolean forceRun = false;
+            final ResponseEntity<JsonNode> extractionResp = extractionProxy.processPdfExtraction(compressImages,
+                    !forceRun,
+                    documentFile);
+
+            final JsonNode body = extractionResp.getBody();
+            final UUID jobId = UUID.fromString(body.get("job_id").asText());
+
+            final int POLLING_INTERVAL_SECONDS = 5;
+            final int MAX_EXECUTION_TIME_SECONDS = 600;
+            final int MAX_ITERATIONS = MAX_EXECUTION_TIME_SECONDS / POLLING_INTERVAL_SECONDS;
+
+            boolean jobDone = false;
+            for (int i = 0; i < MAX_ITERATIONS; i++) {
+
+                final ResponseEntity<JsonNode> statusResp = extractionProxy.status(jobId);
+                if (!statusResp.getStatusCode().is2xxSuccessful()) {
+                    throw new RuntimeException("Unable to poll status endpoint");
+                }
+
+                final JsonNode statusData = statusResp.getBody();
+                if (!statusData.get("error").isNull()) {
+                    throw new RuntimeException("Extraction job failed: " + statusData.has("error"));
+                }
+
+                log.info("Polled status endpoint {} times:\n{}", i + 1, statusData);
+                jobDone = statusData.get("error").asBoolean() || statusData.get("job_completed").asBoolean();
+                if (jobDone) {
+                    break;
+                }
+                Thread.sleep(POLLING_INTERVAL_SECONDS * 1000);
+            }
+
+            if (!jobDone) {
+                throw new RuntimeException("Extraction job did not complete within the expected time");
+            }
+
+            final ResponseEntity<byte[]> zipFileResp = extractionProxy.result(jobId);
+            if (!zipFileResp.getStatusCode().is2xxSuccessful()) {
+                throw new RuntimeException("Unable to fetch the extraction result");
+            }
+
+            final String zipFileName = documentId + "_cosmos.zip";
+            documentService.uploadFile(documentId, zipFileName, new ByteArrayEntity(zipFileResp.getBody()));
+
+            document.getFileNames().add(zipFileName);
+
+            // Open the zipfile and extract the contents
+
+            final Map<String, HttpEntity> fileMap = new HashMap<>();
+            try {
+                final ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(zipFileResp.getBody());
+                final ZipInputStream zipInputStream = new ZipInputStream(byteArrayInputStream);
+
+                ZipEntry entry = zipInputStream.getNextEntry();
+                while (entry != null) {
+
+                    fileMap.put(entry.getName(), zipEntryToHttpEntity(zipInputStream));
+
+                    entry = zipInputStream.getNextEntry();
+                }
+
+                zipInputStream.closeEntry();
+                zipInputStream.close();
+            } catch (final IOException e) {
+                throw new RuntimeException("Unable to extract the contents of the zip file", e);
+            }
+
+            final ResponseEntity<JsonNode> textResp = extractionProxy.text(jobId);
+            if (!textResp.getStatusCode().is2xxSuccessful()) {
+                throw new RuntimeException("Unable to fetch the text extractions");
+            }
+
+            // clear existing assets
+            document.setAssets(new ArrayList<>());
+
+            for (final ExtractionAssetType extractionType : ExtractionAssetType.values()) {
+                final ResponseEntity<JsonNode> response = extractionProxy.extraction(jobId,
+                        extractionType.toStringPlural());
+                log.info(" {} response status: {}", extractionType, response.getStatusCode());
+                if (!response.getStatusCode().is2xxSuccessful()) {
+                    log.warn("Unable to fetch the {} extractions", extractionType);
+                    continue;
+                }
+
+                for (final JsonNode record : response.getBody()) {
+
+                    String fileName = "";
+                    if (record.has("img_pth")) {
+
+                        final String path = record.get("img_pth").asText();
+                        fileName = path.substring(path.lastIndexOf("/") + 1);
+
+                        if (fileMap.containsKey(fileName)) {
+                            log.warn("Unable to find file {} in zipfile", fileName);
+                        }
+
+                        final HttpEntity file = fileMap.get(fileName);
+
+                        documentService.uploadFile(documentId, fileName, file);
+
+                    } else {
+                        log.warn("No img_pth found in record: {}", record);
+                    }
+
+                    final DocumentExtraction extraction = new DocumentExtraction();
+                    extraction.setFileName(fileName);
+                    extraction.setAssetType(extractionType);
+                    extraction.setMetadata(objectMapper.convertValue(record, Map.class));
+
+                    document.getAssets().add(extraction);
+                }
+            }
+
+            String responseText = "";
+            for (final JsonNode record : textResp.getBody()) {
+                if (record.has("content")) {
+                    responseText += record.get("content").asText() + "\n";
+                } else {
+                    log.warn("No content found in record: {}", record);
+                }
+            }
+
+            document.setText(responseText);
+
+            // update the document
+            document = documentService.updateAsset(document).orElseThrow();
+
+            if (document.getText() == null || document.getText().isEmpty()) {
+                log.warn("Document {} has no text to send", documentId);
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Document has no text");
+            }
+
+            // check for input length
+            if (document.getText().length() > 600000) {
+                log.warn("Document {} text too long for GoLLM model card task", documentId);
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Document text is too long");
+            }
+
+            final ModelCardResponseHandler.Input input = new ModelCardResponseHandler.Input();
+            input.setResearchPaper(document.getText());
+
+            // Create the task
+            final TaskRequest req = new TaskRequest();
+            req.setType(TaskRequest.TaskType.GOLLM);
+            req.setScript(ModelCardResponseHandler.NAME);
+            req.setInput(objectMapper.writeValueAsBytes(input));
+
+            final ModelCardResponseHandler.Properties props = new ModelCardResponseHandler.Properties();
+            props.setDocumentId(documentId);
+            req.setAdditionalProperties(props);
+        } catch (final Exception e) {
+            final String error = "Unable to extract pdf";
+            log.error(error, e);
+            throw new ResponseStatusException(
+                    org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR,
+                    error);
+        }
+    }
+
+    public HttpEntity zipEntryToHttpEntity(final ZipInputStream zipInputStream) throws IOException {
+        final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+        final byte[] buffer = new byte[1024];
+        int len;
+        while ((len = zipInputStream.read(buffer)) > 0) {
+            byteArrayOutputStream.write(buffer, 0, len);
+        }
+
+        return new ByteArrayEntity(byteArrayOutputStream.toByteArray());
+    }
+}
